Standardize interface usage and use lambdas
Switches to Consumer<Event> from custom EventHandler class. There is,
as far as I can tell, no advantage to using full static inner classes
rather than lambdas, so I've switched that too. It will be
a runtime error if a lambda accepting the wrong type is passed, but
as long as no one edits the class to break that it should be fine.