Replace redundant Action class with plain Runnable
Actions were really just runnables anyway, and interfaces are preferable
to abstract classes. Additionally, @Danappelxx needs an action which is
also a subsystem manager.

Some changes were made to implements the switch. RunnableAction
is now redundant, but removing it caused a problem with mocking in
the tests, because lambdas are final and cannot be mocked. In order to
fix this problem, I created a new class that implements Runnable in
place of the previous wrapped lambda, and mocked that.