Remove Configuration
Switch from Configuration to constants, either accessed directly
by the class or passed in through the constructor on a per-argument
where there are multiple configurations for the same class.

Use of the Configuration class has caused significant maintainability
problems. The Configuration code itself is complex, but more importantly
it creates problems. Errors that should be visible at compile time are
only visible at runtime, which delays committing code to master until
tests can be run on the robot. This is not a theoretical issue, as
two programers were stuck waiting for electrical work to be done on
the robot for ~30 minutes this morning.

There are minimal advantages to keeping this system. Because the
resource file is packaged into the jar with code, we are unable to make
config modifications without a full deploy. In practice, we are unlikely
to want to maintain multiple sets of constants, because switching robots
would require significant changes to the codebase in any case. In short,
it does little good to keep this code, and there are advantages to its
removal. We are already using an alternate solution (public static final
java constants) for some cases, and migrating fully to this option
allows us to stop splitting our efforts.

There are several parts of the configuration class that could prove
useful in future. The channel conflict monitoring potentially detects
errors (albeit at runtime). It is my opinion that a channel conflict
is fairly unlikely, but this feature can always be reintroduced
separately if needed. The other feature that has been removed is
the dependency injection element, which allows the instantiating class
to control the configuration of a new object. For the time being, this
feature has been removed from cases where there is only one instance
(and thus only one configuration) of class. In the one
alternate case (RobotSide), I switched to explicitly passing in the
config data There is no reason that such classes cannot be shifted to
have configuration injected (on a per argument basis, or using a special
carrier class) in future, and that would probably be preferable to the
Configuration class system anyway. Any separation of concerns problem
is mitigated by the actual constants being stored in a centralized
settings file, so the class is still somewhat insulated from the values
of the constants.

Approved (in general terms) by @Danappelxx