Replace registry with basic dependency injection (#47)
* Replace registry with basic dependency injection

Removes registry and replaces it with dependency injection, from
Robot.java on down through the dependency tree. This should be replaced
with dagger (or another dependency injection framework), but for now
this is sufficient. This change will eliminate runtime failures due to
forgetting to enter an object properly or other registry problems.