Rename interfaces and implementations
Changes interfaces by removing "Interface" and implementations by
adding "Standard".

A person has a phone, not a phoneInterface. An iPhone is a type of
phone: the name of the derivative is derived from the base, not
the other way around. Because nothing other than Robot should be
referencing the implementations, it makes sense that the interfaces
should have the obvious and clear names.